# parser_csv: Improve the performance for typical cases