@repeatedly 2019-08-07
Apply reviews. I will send a patch to fluentd-docs-gitbook.